Tasks #5674: Action menu for group chat screen

+++ app/src/main/java/mega/privacy/android/app/lollipop/adapters/MegaParticipantsChatLollipopAdapter.java
@@ -230,32 +230,39 @@ public class MegaParticipantsChatLollipopAdapter extends RecyclerView.Adapter<Me
 
 				MegaUser contact = megaApi.getContact(((ViewHolderParticipantsList)holder).contactMail);
 
+				if(contact!=null){
+					ChatParticipantAvatarListener listener = new ChatParticipantAvatarListener(context, ((ViewHolderParticipantsList)holder), this);
 
-				ChatParticipantAvatarListener listener = new ChatParticipantAvatarListener(context, ((ViewHolderParticipantsList)holder), this);
-
-				File avatar = null;
-				if (context.getExternalCacheDir() != null) {
-					avatar = new File(context.getExternalCacheDir().getAbsolutePath(), ((ViewHolderParticipantsList)holder).contactMail + ".jpg");
-				} else {
-					avatar = new File(context.getCacheDir().getAbsolutePath(), ((ViewHolderParticipantsList)holder).contactMail + ".jpg");
-				}
-				Bitmap bitmap = null;
-				if (avatar.exists()) {
-					if (avatar.length() > 0) {
-						BitmapFactory.Options bOpts = new BitmapFactory.Options();
-						bOpts.inPurgeable = true;
-						bOpts.inInputShareable = true;
-						bitmap = BitmapFactory.decodeFile(avatar.getAbsolutePath(), bOpts);
-						if (bitmap == null) {
-							avatar.delete();
+					File avatar = null;
+					if (context.getExternalCacheDir() != null) {
+						avatar = new File(context.getExternalCacheDir().getAbsolutePath(), ((ViewHolderParticipantsList)holder).contactMail + ".jpg");
+					} else {
+						avatar = new File(context.getCacheDir().getAbsolutePath(), ((ViewHolderParticipantsList)holder).contactMail + ".jpg");
+					}
+					Bitmap bitmap = null;
+					if (avatar.exists()) {
+						if (avatar.length() > 0) {
+							BitmapFactory.Options bOpts = new BitmapFactory.Options();
+							bOpts.inPurgeable = true;
+							bOpts.inInputShareable = true;
+							bitmap = BitmapFactory.decodeFile(avatar.getAbsolutePath(), bOpts);
+							if (bitmap == null) {
+								avatar.delete();
+								if (context.getExternalCacheDir() != null) {
+									megaApi.getUserAvatar(contact, context.getExternalCacheDir().getAbsolutePath() + "/" + contact.getEmail() + ".jpg", listener);
+								} else {
+									megaApi.getUserAvatar(contact, context.getCacheDir().getAbsolutePath() + "/" + contact.getEmail() + ".jpg", listener);
+								}
+							} else {
+								((ViewHolderParticipantsList)holder).contactInitialLetter.setVisibility(View.GONE);
+								((ViewHolderParticipantsList)holder).imageView.setImageBitmap(bitmap);
+							}
+						} else {
 							if (context.getExternalCacheDir() != null) {
 								megaApi.getUserAvatar(contact, context.getExternalCacheDir().getAbsolutePath() + "/" + contact.getEmail() + ".jpg", listener);
 							} else {
 								megaApi.getUserAvatar(contact, context.getCacheDir().getAbsolutePath() + "/" + contact.getEmail() + ".jpg", listener);
 							}
-						} else {
-							((ViewHolderParticipantsList)holder).contactInitialLetter.setVisibility(View.GONE);
-							((ViewHolderParticipantsList)holder).imageView.setImageBitmap(bitmap);
 						}
 					} else {
 						if (context.getExternalCacheDir() != null) {
@@ -264,15 +271,13 @@ public class MegaParticipantsChatLollipopAdapter extends RecyclerView.Adapter<Me
 							megaApi.getUserAvatar(contact, context.getCacheDir().getAbsolutePath() + "/" + contact.getEmail() + ".jpg", listener);
 						}
 					}
-				} else {
-					if (context.getExternalCacheDir() != null) {
-						megaApi.getUserAvatar(contact, context.getExternalCacheDir().getAbsolutePath() + "/" + contact.getEmail() + ".jpg", listener);
-					} else {
-						megaApi.getUserAvatar(contact, context.getCacheDir().getAbsolutePath() + "/" + contact.getEmail() + ".jpg", listener);
-					}
 				}
+				else{
+					log("Participant is NOT contact");
+				}
+
 			} else {
-				log("Participant is NOT contact");
+				log("NOT email- Participant is NOT contact");
 			}
 
 			int status = participant.getStatus();
